feat: add AND toggle (#4)
Added a toggle in the security limit component that will trigger an AND
query in openseach that will return documents that only contain the
selected releasability. This is done by adding the user's selection as
parameters to the api call similarly to how we add what the user hasn't
selected. Also added a flow tooltip to make it clear what the toggle is
for. Wording should be checked. Depends on
https://github.com/AustralianCyberSecurityCentre/azul-metastore/pull/2